+++ packages/viz/src/modules/print/BasicsPrintView.js
@@ -19,7 +19,8 @@ import _ from 'lodash'
 import i18next from 'i18next'
 import moment from 'moment-timezone'
 
-import PrintView from './PrintView'
+// import PrintView from './PrintView'
+import { PrintView } from 'dumb/src/modules/print/print-view'
 
 import {
   cgmStatusMessage,